requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
langchain>=0.1.0,<0.4.0
langchain-classic>=1.0.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
openai>=1.3.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
openai-whisper>=20231117
edge-tts>=6.1.9
torch>=2.0.0
numpy>=1.24.0
httpx>=0.25.0
python-multipart>=0.0.6
aiofiles>=23.2.1
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydub>=0.25.1
soundfile>=0.12.1
streamlit>=1.28.0
sounddevice>=0.4.6
requests>=2.31.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
